requirements.txt
google-generativeai
google-api-python-client
requests 
beautifulsoup4
dotenv